Now only writes file if new content different from existing content

+++ src/file-writer.js
@@ -11,6 +11,12 @@ module.exports = (rootPath, files) => {
         const fullFilePath = path.join(rootPath, filename)
         const dirname = path.dirname(fullFilePath)
         fs.ensureDirSync(dirname)
-        fs.writeFileSync(fullFilePath, files[filename])
+
+        // TODO: check what happens if file path exists, but is a folder rather than a file.
+
+        // Before writing this content out read the existing content: if nothing's changed, don't write the content out
+        // to avoid unnecessary changes in git, file dates, etc.
+        if (!fs.existsSync(fullFilePath) || fs.readFileSync(fullFilePath) != files[filename])
+            fs.writeFileSync(fullFilePath, files[filename])
     })
 }
